@@ -48,11 +48,12 @@
     "postcss-loader": "3.0.0",
     "postcss-preset-env": "6.0.6",
     "postcss-safe-parser": "4.0.1",
-    "react": "^16.6.1",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
     "react-app-polyfill": "^0.1.3",
     "react-cookie": "^3.0.7",
     "react-dev-utils": "^6.1.1",
-    "react-dom": "^16.6.1",
+    "react-dom": "^16.8.6",
     "react-redux": "^5.1.1",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
