@@ -1,20 +1,24 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import styles from './styles.module.css';
 
-const SidebarComponent = (props) => {
-	if (props.path[1] !== undefined && props.path[1] === props.id){
-		return (
-			<div className={styles.container}>
+
+function SidebarComponent(props) {
+	const [style, setStyle] = useState(styles.triangle_none);
+	useEffect(() => {
+		setStyle(styles.triangle_none);
+		if (props.path[1] == props.id) {
+			setStyle(styles.triangle);
+		}
+	}, props.path[1]);
+
+
+	return(
+		<div className={styles.container}>
 			<p className={styles.SidebarComponent_active}>{props.topic}{props.name}</p> 
-			<div className={styles.triangle}></div>
-			</div>
-		)
-	}
-	else {
-		return (
-			<p className={styles.SidebarComponent}>{props.topic}{props.name}</p> 
-		)
-	}		
-};
+			<div className={style}></div>
+		</div>
+	);
+}
+
 
 export default SidebarComponent;
